TH fixes for ghc-8.10.1
Due to GHC issue https://gitlab.haskell.org/ghc/ghc/-/issues/16881,
Template Haskell behaves differently in ghc-8.10.1 from previous
versions in that tuples of size one are "explicitly" created and are
different from the underlying type.

However, the current TH code in Marlowe assumes the old behaviour.
This is a compatibility fix to make the code work both in the old
and in the new setting.