Generalize `Contract` over identity and token
This introduces two type parameters `i` and `t` that generalize over the
type of identity (`ByteString` in `marlowe`, `PubKeyHash` in
`marlowe-cardano`) and token (`Token` by default). Nothing in the
language definition actually cares what these two types are.

For the main definitions in each module, I've introduced a new
"generalized" function (e.g., `computeTransaction'`), and then defined
the "old" function (e.g., `computeTransaction`) in terms of that more
general one.

I have also generalized all the examples; the examples don't care what
`i` and `t` are either, _except_ that they often hardcode a choice of
`ada` as the token; but this doesn't change the examples at all. So for
the examples I've added a separate token type as an additional argument.
(The examples always use roles, so they really are polymorphic in `i`.)